refactor: move the findRecommendedFilms method
from the UserService class to the FilmService class
