+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -5,12 +5,16 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
+import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.director.DirectorDaoStorage;
 import ru.yandex.practicum.filmorate.storage.film.FilmDaoStorage;
 import ru.yandex.practicum.filmorate.storage.genre.GenreDaoStorage;
 import ru.yandex.practicum.filmorate.storage.like.LikeDaoStorage;
+import ru.yandex.practicum.filmorate.storage.user.UserDaoStorage;
 
 import java.util.ArrayList;
+import java.util.Comparator;
+import java.util.HashMap;
 import java.util.List;
 
 @Service
@@ -22,6 +26,7 @@ public class FilmService {
     private final LikeDaoStorage likeDaoStorage;
     private final GenreDaoStorage genreDaoStorage;
     private final DirectorDaoStorage directorDaoStorage;
+    private final UserDaoStorage userDaoStorage;
 
     private void loadData(Film film) {
         film.setGenres(genreDaoStorage.getGenresByFilm(film));
@@ -104,5 +109,38 @@ public class FilmService {
 
     public List<Film> getCommonFilms(Long userId, Long friendId) {
         return filmDaoStorage.getCommonFilms(userId, friendId);
+    }
+
+    public List<Film> findRecommendedFilms(Long id) {
+        HashMap<User, List<Film>> filmsTable = new HashMap<>();
+        List<Film> userFilms = filmDaoStorage.findFilmsLikedByUser(id);
+        List<User> users = userDaoStorage.getAllUser();
+
+        users.remove(userDaoStorage.getUserById(id));
+
+        for (User other : users) {
+            List<Film> otherFilms = filmDaoStorage.findFilmsLikedByUser(other.getId());
+            filmsTable.put(other, otherFilms);
+        }
+
+        List<List<Film>> differencesTable = new ArrayList<>();
+
+        for (List<Film> value : filmsTable.values()) {
+            List<Film> filmsPackage = new ArrayList<>();
+
+            for (Film film : value) {
+                film = getFilmById(film.getId());
+                if (!userFilms.contains(film)) {
+                    filmsPackage.add(film);
+                }
+            }
+            differencesTable.add(filmsPackage);
+        }
+
+        differencesTable.removeIf(List::isEmpty);
+
+        return differencesTable.stream()
+                .min(Comparator.comparing(List<Film>::size))
+                .orElse(new ArrayList<>());
     }
 }
